fixed behaviour with req body, extended log middleware
