+++ nodes/node_balancer/cmd/blockchain.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"fmt"
 	"log"
 	"strconv"
 
@@ -9,7 +10,51 @@ import (
 
 var (
 	nodeConfigs NodeConfigs
+
+	ALLOWED_METHODS = []string{
+		"eth_blockNumber",
+		"eth_call",
+		"eth_chainId",
+		"eth_estimateGas",
+		"eth_feeHistory",
+		"eth_gasPrice",
+		"eth_getBalance",
+		"eth_getBlockByHash",
+		"eth_getBlockByNumber",
+		"eth_getBlockTransactionCountByHash",
+		"eth_getBlockTransactionCountByNumber",
+		"eth_getCode",
+		"eth_getLogs",
+		"eth_getStorageAt",
+		"eth_getTransactionByHash",
+		"eth_getTransactionByBlockHashAndIndex",
+		"eth_getTransactionByBlockNumberAndIndex",
+		"eth_getTransactionCount",
+		"eth_getTransactionReceipt",
+		"eth_getUncleByBlockHashAndIndex",
+		"eth_getUncleByBlockNumberAndIndex",
+		"eth_getUncleCountByBlockHash",
+		"eth_getUncleCountByBlockNumber",
+		"eth_getWork",
+		"eth_mining",
+		// "eth_sendRawTransaction",
+		"eth_protocolVersion",
+		"eth_syncing",
+
+		"net_listening",
+		"net_peerCount",
+		"net_version",
+
+		"web3_clientVersion",
+	}
 )
+
+type JSONRPCRequest struct {
+	Jsonrpc string        `json:"jsonrpc"`
+	Method  string        `json:"method"`
+	Params  []interface{} `json:"params"`
+	ID      uint64        `json:"id"`
+}
 
 // Node conf
 type BlockchainConfig struct {
@@ -26,6 +71,15 @@ type NodeConfig struct {
 
 type NodeConfigs struct {
 	NodeConfigs []NodeConfig
+}
+
+func verifyMethodWhitelisted(method string) error {
+	for _, m := range ALLOWED_METHODS {
+		if method == m {
+			return nil
+		}
+	}
+	return fmt.Errorf("Method not allowed")
 }
 
 // Return list of NodeConfig structures
